Adding Publications page
Adding a publications page, including links to it in the navigation bar
on all pages.  Also added a bib subdirectory that includes a .bib file I
exported from my Mendeley database with the CISM-related references I
could find.  This dir also includes a readme about how to convert that
bib file into html to be pasted into the publications directory.